Fix Double to Int32 conversion error in TextExportManager
- Fixed line 144 where Double quantity was passed to function expecting Int32
- Added Int32(quantity) conversion for HoldingGainLossCalculator.calculateGainLossWithRawValues call
- TextExportManager.swift now compiles without type conversion errors

+++ MyAsset/MyAsset/TextExportManager.swift
@@ -141,7 +141,7 @@ class TextExportManager: ObservableObject {
         let rawGainLoss = HoldingGainLossCalculator.calculateGainLossWithRawValues(
             currentPrice: currentPrice,
             purchasePrice: purchasePrice, 
-            quantity: quantity,
+            quantity: Int32(quantity),
             symbol: symbol,
             context: "EXPORT-RAW"
         )
